Move trade header into trade component
