+++ src/pages/Trade/Trade.module.css
@@ -465,8 +465,7 @@ background-size: 300% 100%;
         width: 100%;
         height: 100%;
         overflow: auto;
-        padding: 8px;
-        padding-bottom: 0px;
+        padding: 0px 8px;
     }
     .main__chart_container::-webkit-scrollbar {
         display: none;
