+++ src/pages/Trade/TradeCharts/TradeCharts.module.css
@@ -33,7 +33,6 @@
 .tutorial_button:hover {
     color: var(--text1);
 }
-.token_info_container,
 .time_frame_container {
     display: flex;
     flex-direction: row;
@@ -89,14 +88,6 @@
     0px 0px 12px rgba(205, 193, 255, 0.2), 0px 0px 7px rgba(205, 193, 255, 0.2),
     0px 0px 4px var(--accent5), 0px 0px 2px rgba(205, 193, 255, 0.2); */
 }
-.fullscreen_button {
-    outline: none;
-    border: none;
-    background: transparent;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
 
 .active_selected_button {
     width: 100%;
@@ -106,43 +97,6 @@
 
 .non_active_selected_button {
     color: var(--text2);
-}
-
-.token_info_container .amount {
-    font-size: var(--header2-size);
-    line-height: var(--header2-lh);
-    color: var(--text2);
-}
-
-.token_info_container .change_positive {
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    color: var(--positive);
-}
-
-.token_info_container .change_negative {
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    color: var(--negative);
-}
-
-.token_info_container .tvl_display {
-    color: var(--text2);
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-}
-
-.graph_settings_container {
-    width: 100%;
-    /* max-height: 100%; */
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-    align-items: center;
-    gap: 4px; /*     gap: 1rem;; */
-
-    display: none;
-    display: none;
 }
 
 .chart_item_container {
@@ -326,8 +280,7 @@
     }
 
     /* .chart_image,
-    .time_frame_container,
-    .token_info_container {
+    .time_frame_container {
         display: none;
     } */
 }
@@ -357,18 +310,5 @@
 
     .expand_table {
         height: 100%;
-    }
-
-    .graph_settings_container {
-        width: 100%;
-        display: flex;
-        flex-direction: row;
-        justify-content: flex-end;
-        align-items: center;
-        gap: 1rem;
-    }
-
-    .graph_settings_container svg {
-        cursor: pointer;
     }
 }
