+++ src/pages/Trade/TradeCharts/TradeChartsComponents/TradeChartsTokenInfo.module.css
@@ -96,8 +96,7 @@
     }
 
     .chart_image,
-    .time_frame_container,
-    .token_info_container {
+    .time_frame_container {
         display: none;
     }
 }
